Move convertion to its own file
Move convertion to its own file
